remove RowValue component, add rowValue prop instead. Fix tags column logic
